requirements.txt
connexion[swagger-ui, uvicorn]>=3.2.0
sqlalchemy[asyncio]>=2.0.0
marshmallow-sqlalchemy>=0.28.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-aiohttp>=1.0.0
aiosqlite>=0.21.0
aiohttp>=3.8.0
pytz>=2025.2